+++ DependencyInjection/Configuration.php
@@ -51,10 +51,36 @@ final class Configuration implements ConfigurationInterface
                     ->end()
                 ->end()
                 ->arrayNode('throttle')
+                    ->beforeNormalization()
+                        ->ifTrue(function ($v) { return is_array($v) && (isset($v['limit']) || isset($v['period'])); })
+                        ->then(function ($v) {
+                            $v['default'] = [];
+                            if (isset($v['limit'])) {
+                                @trigger_error('The indragunawan_api_rate_limit.throttle.limit configuration key is deprecated since version v0.2.0 and will be removed in v0.3.0. Use the indragunawan_api_rate_limit.throttle.default.limit configuration key instead.', E_USER_DEPRECATED);
+
+                                $v['default']['limit'] = $v['limit'];
+                            }
+
+                            if (isset($v['period'])) {
+                                @trigger_error('The indragunawan_api_rate_limit.throttle.period configuration key is deprecated since version v0.2.0 and will be removed in v0.3.0. Use the indragunawan_api_rate_limit.throttle.default.period configuration key instead.', E_USER_DEPRECATED);
+
+                                $v['default']['period'] = $v['period'];
+                            }
+
+                            return $v;
+                        })
+                    ->end()
                     ->addDefaultsIfNotSet()
                     ->children()
                         ->integerNode('limit')->defaultValue(60)->end()
                         ->integerNode('period')->min(1)->defaultValue(60)->end()
+                        ->arrayNode('default')
+                            ->addDefaultsIfNotSet()
+                            ->children()
+                                ->integerNode('limit')->defaultValue(60)->end()
+                                ->integerNode('period')->min(1)->defaultValue(60)->end()
+                            ->end()
+                        ->end()
                     ->end()
                 ->end()
                 ->arrayNode('exception')