+++ CHANGELOG.md
@@ -5,3 +5,4 @@ master
 ------
 
 * deprecated the `indragunawan_api_rate_limit.storage` configuration key.
+* deprecated the `indragunawan_api_rate_limit.throttle.period` and `indragunawan_api_rate_limit.throttle.limit` configuration key.